add ga scheme

+++ include/evolution/population/individual.h
@@ -5,7 +5,7 @@
 class Individual{
 public:
     int skill_factor;
-    int central_id;
+    unsigned long central_id;
     std::vector <float> objectives;
     Tree* genes;
     Individual(Tree* genes, int skill_factor);
